refactor config + add test
